refactoring: Modal과 Infromation병합하기
