+++ src/components/Card/CardFooter.tsx
@@ -1,10 +1,11 @@
 import { PropsWithChildren } from 'react';
+
 import * as S from './index.style';
 
 interface ModalFooterProps extends PropsWithChildren {}
 
 const ModalFooter = ({ children }: ModalFooterProps) => {
-  return <S.ModalFooter>{children}</S.ModalFooter>;
+  return <S.CardFooter>{children}</S.CardFooter>;
 };
 
 export default ModalFooter;
