+++ src/components/Card/index.style.ts
@@ -21,4 +21,15 @@ export const CardHeader = styled.div`
 export const CardBody = styled.div`
   width: 100%;
   padding: 10px 5px;
+
+  display: flex;
+  flex-direction: column;
+`;
+
+export const CardFooter = styled.div`
+  width: 100%;
+  padding: 10px 5px;
+
+  display: flex;
+  flex-direction: column;
 `;
