--- src/components/Information/index.style.ts
@@ -1,24 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Information = styled.div`
-  width: 100%;
-  padding: 10px 20px;
-
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-`;
-
-export const InformationHeader = styled.div`
-  width: 100%;
-
-  font-size: 20px;
-
-  display: flex;
-  flex-direction: column;
-`;
-
-export const InformationBody = styled.div`
-  width: 100%;
-  padding: 10px 5px;
-`;
